refactor: useSingleton
